Add MeterAdapter support for .NET 8 static meter/instrument tags.
These tags are set on the meter and/or instrument at creation time, and are automatically added to the measurements' labels.

This required a little bit of internal reworking of MeterAdapter so that the static label values are stored into the metric context. I re-organized the code of OnMeasurementRecorded to re-use code between the metric types better, as I had to move the label value pool fetch to be gotten AFTER the metric context.

Added a target framework for .NET 8, since we obviously need it for the new feature. I multi-targeted the tests to both .NET 7 and .NET 8 to make sure I didn't break anything.

Added these to the example and wrote tests for them too.